refactor: convert the entire project from a composite build to a build with sub-projects, as the IDE friction was getting too much.

+++ build.sh
@@ -1,6 +1,9 @@
 #!/bin/sh
 
-rm -rf special-models/build special-models/.gradle limlib/build limlib/.gradle build .gradle $HOME/.gradle
-cd special-models; ./gradlew --no-daemon --quiet :clean :build; cd -; 
-cd limlib; ./gradlew --no-daemon --quiet :build; cd -; 
+if [ "$1" -eq "-f" ]; then
+    rm -rf special-models/build special-models/.gradle limlib/build limlib/.gradle build .gradle $HOME/.gradle
+fi
+
+# cd special-models; ./gradlew --no-daemon --quiet :clean :build; cd -; 
+# cd limlib; ./gradlew --no-daemon --quiet :build; cd -; 
 ./gradlew --no-daemon --quiet :build
